remove db/schema.rb from source control
@benjamin-sweeney even though the rails core team recommends keeping
db/schema.rb under source control I will be removing it since
it seems to suite our team better.

The prod database is PG and I personally use PG locally,
but I changed the default local DB to sqlite3 (5cc5c0188c1e7f789c) since it is far
easier to setup, and everyone was having enough dificulties as it
is to get that to work. With different people using different DBs we
are going to see schema.rb conflicts. Better to just gitignore it
for now.